refactor: type folder & imports
