+++ src/components/CurrentVehicle.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { ConfigData } from "../utils/types";
+import type { ConfigData } from "../types";
 
 interface CurrentVehicleProps {
   currVehicle: ConfigData | null;
